styles n stuff
break organizationDetails up into context and DOM elements

platformicon components
