+++ src/sentry/static/sentry/app/views/onboarding/utils.js
@@ -1,4 +1,20 @@
-const onboardingSteps = {organization: 0, project: 1, configure: 2};
-// const onboardingStepDescriptions = {project: 0, configure: 1, next: 2};
+import {platforms} from '../../../../../integration-docs/_platforms.json';
 
-export {onboardingSteps};
+const onboardingSteps = {organization: 0, project: 1, configure: 2};
+
+const stepDescriptions = {
+  organization: 'Create an Organization in Sentry',
+  project: 'Tell us about your project',
+  configure: 'Configure your application and send an event'
+};
+
+const flattenedPlatforms = [].concat(
+  [],
+  ...platforms.map(language => {
+    return language.integrations.map(i => {
+      return {...i, language: language.id};
+    });
+  })
+);
+
+export {onboardingSteps, stepDescriptions, flattenedPlatforms};
