+++ src/sentry/static/sentry/less/shared-components.less
@@ -3351,7 +3351,7 @@ ul.tag-list {
 
   // Colors
 
-  li {
+  li, span {
     &.python .platformicon {
       background: darken(@blue, 6);
     }
