+++ src/sentry/web/urls.py
@@ -257,7 +257,7 @@ urlpatterns += patterns(
     url(r'^account/settings/social/', include('social_auth.urls')),
     url(r'^account/', generic_react_page_view),
 
-    # url(r'^onboarding/', generic_react_page_view),
+    url(r'^onboarding/', generic_react_page_view),
 
     # Admin
     url(r'^manage/status/environment/$', admin.status_env, name='sentry-admin-status'),
